requirements.txt
beautifulsoup4
chromadb
crewai
docx2txt
googlesearch-python
langchain_chroma
langchain_community
langchain_core
langchain_huggingface
mistralai
openpyxl
pandas
pypdf
python-dotenv
requests
sentence-transformers